Currencies

+++ app/src/main/java/pl/mazak/cigscanner/data/db/currency/Currency.kt
@@ -2,12 +2,11 @@ package pl.mazak.cigscanner.data.db.currency
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.math.BigDecimal
 
 @Entity(tableName = "currency")
 data class Currency(
     @PrimaryKey
     val id: Int,
-    val euro: BigDecimal,
-    val czechCrown: BigDecimal
+    val euro: Double,
+    val czechCrown: Double
 )
